Add files via upload
Components Needed:
Arduino Uno
GPS Module (e.g., NEO-6M)
Motor Shield (e.g., L298N)
Accelerometer (for detecting accidents)
Mobile App (for remote control or notifications)
Power Supply and necessary cables
Overview of Steps:
1. Set up the GPS Module:
Connect the GPS module to the Arduino Uno using the appropriate pins (usually TX, RX, VCC, GND).

Write a program to interface with the GPS module. This program will read GPS data (latitude, longitude, speed, etc.) from the module.

2. Add an Accelerometer:
Connect an accelerometer (e.g., MPU-6050) to the Arduino Uno. This will be used to detect sudden changes in acceleration which may indicate an accident.

Write a program to read data from the accelerometer and process it to detect potential accidents.

3. Control Motors:
Connect the motors to the motor shield. The motor shield will interface with the Arduino Uno and allow you to control the motors.

Write a program that can control the motors based on input from the accelerometer or other sensors. For example, it could apply emergency braking if an accident is detected.

4. Integrate GPS and Accelerometer Data:
Combine the data from the GPS module and accelerometer to make informed decisions about potential accidents.

For example, if the accelerometer detects a sudden stop, check the GPS data to see if the vehicle has indeed stopped or if it's a false positive.

5. Implement Accident Detection Algorithm:
Based on the combined data, develop an algorithm to detect accidents. This could involve checking for sudden deceleration, a sharp change in direction, or other criteria that might indicate an accident.

6. Add Connectivity (Optional):
If you want to control or monitor the system remotely, you can add connectivity. You might use Bluetooth or Wi-Fi for communication with a mobile app.

Develop a mobile app (for Android or iOS) to communicate with the Arduino. The app might allow for remote control or send notifications in case of an accident.

7. Test and Iterate:
Thoroughly test the system to ensure it works as expected. Make any necessary adjustments or improvements.